// mollys-movies-ui/src/app/api/model/movie-download-status-code.ts
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum MovieDownloadStatusCode {
    Started = 'Started',
    Downloaded = 'Downloaded',
    Complete = 'Complete'
};